+++ cmd/logger.go
@@ -2,7 +2,9 @@ package cmd
 
 import (
 	"fmt"
+	"net/url"
 	"os"
+	"runtime"
 	"time"
 
 	"code.vegaprotocol.io/shared/paths"
@@ -56,12 +58,12 @@ func DefaultConfig() zap.Config {
 	}
 }
 
-func BuildJSONLogger(level string, logsDir paths.StatePath) (*zap.Logger, error) {
+func BuildJSONLogger(level string, vegaPaths paths.Paths, logsDir paths.StatePath) (*zap.Logger, string, error) {
 	cfg := DefaultConfig()
 
 	l, err := getLevel(level)
 	if err != nil {
-		return nil, err
+		return nil, "", err
 	}
 
 	cfg.Level = zap.NewAtomicLevelAt(*l)
@@ -71,19 +73,28 @@ func BuildJSONLogger(level string, logsDir paths.StatePath) (*zap.Logger, error)
 	pathSuffix := fmt.Sprintf("%d-%s.log", pid, date)
 
 	logFile := paths.JoinStatePath(logsDir, pathSuffix)
-	appLogPath, err := paths.CreateDefaultStatePathFor(logFile)
+	appLogPath, err := vegaPaths.CreateStatePathFor(logFile)
 	if err != nil {
-		return nil, fmt.Errorf("failed getting path for %s: %w", logFile, err)
+		return nil, "", fmt.Errorf("failed getting path for %s: %w", logFile, err)
 	}
 
-	cfg.OutputPaths = []string{appLogPath}
-	cfg.ErrorOutputPaths = []string{appLogPath}
+	zapLogPath := appLogPath
+	if runtime.GOOS == "windows" {
+		err := zap.RegisterSink("winfile", newWinFileSink)
+		if err != nil {
+			return nil, "", fmt.Errorf("couldn't register the windows file sink: %w", err)
+		}
+		zapLogPath = "winfile:///" + appLogPath
+	}
+
+	cfg.OutputPaths = []string{zapLogPath}
+	cfg.ErrorOutputPaths = []string{zapLogPath}
 
 	log, err := cfg.Build()
 	if err != nil {
-		return nil, fmt.Errorf("couldn't create logger: %w", err)
+		return nil, "", fmt.Errorf("couldn't create logger: %w", err)
 	}
-	return log, nil
+	return log, appLogPath, nil
 }
 
 func BuildLogger(output, level string) (*zap.Logger, error) {
@@ -147,4 +158,13 @@ func isSupportedLogLevel(level string) bool {
 		}
 	}
 	return false
+}
+
+// newWinFileSink creates a log sink on Windows machines as zap, by default,
+// doesn't support Windows paths. A workaround is to create a fake winfile
+// scheme and register it with zap instead. This workaround is taken from
+// the GitHub issue at https://github.com/uber-go/zap/issues/621.
+func newWinFileSink(u *url.URL) (zap.Sink, error) {
+	// Remove leading slash left by url.Parse().
+	return os.OpenFile(u.Path[1:], os.O_WRONLY|os.O_APPEND|os.O_CREATE, 0o644) // nolint:gomnd
 }